requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
